Use tmpdirs for running tests
The main reason for these changes is to ensure that all functional
testing is done in tmpdirs. Doing this avoids any issues with clean
up, preventing random files from appearing (such as when tests fail).

The .conf files used for the tests are now generated progamatically
from within the test (also into the tmpdir), to make the tests easier
to read, and to allow more flexibility (such as pointing them to the
pdf assets in the tmpdir).

All tests now use absolute paths, the use of relative paths is
not specifically tested. Since all path handling is done using
core library components in standard ways, I don't expect this to be
an issue.