Use appveyor for testing on Windows
For better cross-platform support, CI tests should be done on Windows as well as Linux. Whilst Travis doesn't support Windows testing, appveyor.com is a similar service that only does Windows testing.

As well as setting up the testing, some other changes were necessary to make all the tests pass. I accidentally based my changes on the dev branch instead of master, so have lost some of the history as well as leaking a couple of dev changes in with this commit. They were going to get added anyway, it just makes the history (and this commit) a bit messier than it should be.

Also note that installing dependencies on appveyor isn't quite ideal at the moment, I'd like to get them all packaged on chocolatey and installing properly from there, but for now this works.