Improve testing of ghostscript
This starts to improve the testing of the ghostscript module. Where possible OS specific tests are mocked to keep cross-platform test coverage high, however in some places it is not possible to mock os-specific elements, so these are tested seperately on windows or linux.
Some tidying has been done within the gs module to make testing easier too.